requirements.txt
django>=3.1
djangorestframework>=3.12
django-filter>=2.4.0
django-cors-headers>=3.7.0
drf-extensions>=0.7.0
requests>=2.25.1
pytest>=6.2.3
gunicorn>=20.1.0
django-heroku